refactor: flatten monorepo and migrate to Bun runtime
Major architectural refactor:
- Removed packages/agents dependency by inlining necessary utilities
- Flattened packages/ingester → ingesters/ at repository root
- Migrated from Node.js/pnpm/TypeScript compilation to Bun runtime
- Added robust path resolution system (utils/paths.ts) using .git detection
- Updated IngesterFactory to use static imports for Bun compatibility
- Simplified Docker configuration for Bun-based execution
- Updated all documentation to reflect new architecture

Breaking changes:
- No longer using pnpm workspace or Turborepo
- Ingester now runs TypeScript directly via Bun (no dist/ directory)
- All paths are now relative to repository root via path utilities

Benefits:
- Simpler architecture (no monorepo overhead)
- Faster test execution (~140ms for 92 tests)
- No build step required for ingester
- More maintainable with fewer dependencies
- Clearer separation between Python backend and TypeScript ingester
