--- .npmrc
@@ -1,4 +1,0 @@
-registry=https://registry.npmjs.org/
-network-timeout=600000
-cache=/tmp/.npm-cache
-strict-ssl=false
