+++ ingesters/package.json
@@ -1,11 +1,10 @@
 {
-  "name": "@cairo-coder/ingester",
+  "name": "cairo-coder-ingester",
   "version": "1.0.0",
   "scripts": {
-    "test": "jest",
-    "build": "tsc -p tsconfig.json",
-    "generate-embeddings": "node ./dist/src/generateEmbeddings.js",
-    "generate-embeddings:yes": "node ./dist/src/generateEmbeddings.js -y"
+    "test": "bun test",
+    "generate-embeddings": "bun run src/generateEmbeddings.ts",
+    "generate-embeddings:yes": "bun run src/generateEmbeddings.ts -y"
   },
   "dependencies": {
     "@antora/lunr-extension": "1.0.0-alpha.10",
@@ -23,13 +22,9 @@
     "winston": "^3.17.0"
   },
   "devDependencies": {
-    "@cairo-coder/typescript-config": "workspace:*",
     "@types/jest": "^29.5.14",
-    "jest": "^29.7.0",
-    "nodemon": "^3.1.9",
+    "@types/node": "^20",
     "prettier": "^3.5.2",
-    "ts-jest": "^29.2.5",
-    "ts-node": "^10.9.2",
     "typescript": "^5.7.3"
   }
 }
