+++ ingesters/src/BaseIngester.ts
