+++ ingesters/src/ingesters/StarknetDocsIngester.ts
@@ -50,7 +50,8 @@ export class StarknetDocsIngester extends MarkdownIngester {
    * @returns string - Path to the extract directory
    */
   protected getExtractDir(): string {
-    return path.join(__dirname, '..', '..', 'temp', 'starknet-docs');
+    const { getTempDir } = require('../utils/paths');
+    return getTempDir('starknet-docs');
   }
 
   /**
