+++ ingesters/src/shared.ts
