+++ ingesters/src/utils/contentUtils.ts
