+++ ingesters/src/utils/fileUtils.ts
