+++ ingesters/src/utils/vectorStoreUtils.ts
