--- package.json
@@ -1,27 +1,0 @@
-{
-  "name": "cairo-coder",
-  "private": true,
-  "version": "0.2.2",
-  "scripts": {
-    "build": "turbo run build",
-    "lint": "turbo run lint",
-    "generate-embeddings": "turbo run generate-embeddings",
-    "generate-embeddings:yes": "turbo run generate-embeddings:yes",
-    "clean": "find packages -type d -name 'dist' -exec rm -rf {} +; find packages -type d -name '.turbo' -exec rm -rf {} +",
-    "clean:all": "npm run clean && find packages -type d -name 'node_modules' -exec rm -rf {} + && rm -rf node_modules && pnpm clean"
-  },
-  "devDependencies": {
-    "turbo": "latest",
-    "typescript": "^5.7.2"
-  },
-  "engines": {
-    "node": ">=20",
-    "pnpm": ">=9"
-  },
-  "packageManager": "pnpm@9.10.0",
-  "pnpm": {
-    "overrides": {
-      "error-ex": "<=1.3.2"
-    }
-  }
-}
