--- packages/ingester/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "@cairo-coder/typescript-config/base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "allowSyntheticDefaultImports": true,
-    "skipLibCheck": true,
-    "skipDefaultLibCheck": true
-  },
-  "include": [
-    "src",
-    "**/*.test.ts",
-    "asciidoc/lunr-tokenizer.js",
-    "asciidoc/asciidoc-output-extension.js"
-  ],
-  "exclude": ["node_modules", "**/*.spec.ts"]
-}
