--- packages/typescript-config/base.json
@@ -1,16 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Default",
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "target": "es2022",
-    "allowJs": true,
-    "resolveJsonModule": true,
-    "moduleDetection": "force",
-    "isolatedModules": true,
-    "noUncheckedIndexedAccess": true,
-    "module": "NodeNext"
-  },
-  "exclude": ["node_modules", "dist"]
-}
