--- packages/typescript-config/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "@cairo-coder/typescript-config"
-}
