--- turbo.json
@@ -1,65 +1,0 @@
-{
-  "$schema": "https://turborepo.org/schema.json",
-  "globalDependencies": [".env"],
-  "tasks": {
-    "topo": {
-      "dependsOn": ["^topo"]
-    },
-    "check-types": {
-      "dependsOn": ["topo"]
-    },
-    "build": {
-      "dependsOn": ["^build"],
-      "outputs": ["dist/**", "build/**", "out/**", ".next/**"]
-    },
-    "lint": {},
-    "test": {
-      "dependsOn": ["^build"],
-      "outputs": []
-    },
-    "generate-embeddings": {
-      "dependsOn": ["build"],
-      "cache": false,
-      "interactive": true,
-      "persistent": true,
-      "env": [
-        "POSTGRES_HOST",
-        "POSTGRES_PORT",
-        "POSTGRES_USER",
-        "POSTGRES_PASSWORD",
-        "POSTGRES_DB",
-        "POSTGRES_TABLE_NAME",
-        "OPENAI_API_KEY",
-        "ANTHROPIC_API_KEY",
-        "GEMINI_API_KEY",
-        "DEFAULT_CHAT_PROVIDER",
-        "DEFAULT_CHAT_MODEL",
-        "DEFAULT_EMBEDDING_PROVIDER",
-        "DEFAULT_EMBEDDING_MODEL"
-      ]
-    },
-    "generate-embeddings:yes": {
-      "dependsOn": ["build"],
-      "inputs": ["dist/**", "build/**", "out/**", ".next/**", ".env"],
-      "cache": false,
-      "interactive": false,
-      "persistent": true,
-      "env": [
-        "POSTGRES_HOST",
-        "POSTGRES_PORT",
-        "POSTGRES_USER",
-        "POSTGRES_PASSWORD",
-        "POSTGRES_DB",
-        "POSTGRES_TABLE_NAME",
-        "OPENAI_API_KEY",
-        "ANTHROPIC_API_KEY",
-        "GEMINI_API_KEY",
-        "DEFAULT_CHAT_PROVIDER",
-        "DEFAULT_CHAT_MODEL",
-        "DEFAULT_EMBEDDING_PROVIDER",
-        "DEFAULT_EMBEDDING_MODEL"
-      ]
-    }
-  },
-  "ui": "tui"
-}
